move detail table tests
