+++ api-js/src/types/base/dmarc-report/detail-tables/__tests__/dmarc-failure-table.test.js
@@ -1,22 +1,22 @@
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { graphql, GraphQLSchema } = require('graphql')
-const { makeMigrations } = require('../../migrations')
-const { createQuerySchema } = require('../queries')
-const { createMutationSchema } = require('../mutations')
+const { makeMigrations } = require('../../../../../../migrations')
+const { createQuerySchema } = require('../../../../../queries')
+const { createMutationSchema } = require('../../../../../mutations')
 const bcrypt = require('bcrypt')
 
-const { cleanseInput } = require('../validators')
+const { cleanseInput } = require('../../../../../validators')
 const {
   checkDomainPermission,
   checkDomainOwnership,
   tokenize,
   userRequired,
-} = require('../auth')
+} = require('../../../../../auth')
 const {
   domainLoaderByDomain,
   userLoaderByUserName,
   userLoaderByKey,
-} = require('../loaders')
+} = require('../../../../../loaders')
 const { DB_PASS: rootPass, DB_URL: url } = process.env
 
 describe('given findDomainByDomain query', () => {
@@ -126,7 +126,7 @@ describe('given findDomainByDomain query', () => {
     await drop()
   })
 
-  describe('find the dmarc report full pass information', () => {
+  describe('find the dmarc report dmarc fail information', () => {
     let user
     beforeEach(async () => {
       const userCursor = await query`
@@ -153,31 +153,32 @@ describe('given findDomainByDomain query', () => {
           REMOVE affiliation IN affiliations
       `
     })
-    it('returns full pass data', async () => {
+    it('returns dmarc fail data', async () => {
       const dmarcReportLoader = jest.fn().mockReturnValue({
         data: {
           dmarcSummaryByPeriod: {
             detailTables: {
-              fullPass: {
+              dmarcFailure: {
                 edges: [
                   {
-                    cursor: 'ZnVsbFBhc3M6MQ==',
+                    cursor: 'ZmFpbERtYXJjOjE=',
                     node: {
-                      dkimDomains: 'test.gc.ca',
-                      dkimSelectors: 'selector1',
-                      dnsHost: 'test.canada.ca',
-                      envelopeFrom: 'test.gc.ca',
-                      headerFrom: 'test.gc.ca',
-                      id: 'ZnVsbFBhc3M6MQ==',
+                      disposition: 'none',
+                      dkimDomains: '',
+                      dkimSelectors: '',
+                      dnsHost: 'test.dns.gc.ca',
+                      envelopeFrom: 'test.domain.canada.ca',
+                      headerFrom: 'test.domain.canada.ca',
+                      id: 'ZmFpbERtYXJjOjE=',
                       sourceIpAddress: '123.456.78.91',
-                      spfDomains: 'test.gc.ca',
-                      totalMessages: 50,
+                      spfDomains: 'test.domain.gc.ca',
+                      totalMessages: 30,
                     },
                   },
                 ],
                 pageInfo: {
-                  startCursor: 'ZnVsbFBhc3M6MQ==',
-                  endCursor: 'ZnVsbFBhc3M6MQ==',
+                  startCursor: 'ZmFpbERtYXJjOjE=',
+                  endCursor: 'ZmFpbERtYXJjOjE=',
                   hasNextPage: true,
                   hasPreviousPage: false,
                 },
@@ -194,13 +195,14 @@ describe('given findDomainByDomain query', () => {
             findDomainByDomain(domain: "test.gc.ca") {
               dmarcSummaryByPeriod(month: SEPTEMBER, year: "2020") {
                 detailTables {
-                  fullPass(first: 1) {
+                  dmarcFailure(first: 1) {
                     edges {
                       cursor
                       node {
                         id
                         dkimDomains
                         dkimSelectors
+                        disposition
                         dnsHost
                         envelopeFrom
                         headerFrom
@@ -256,26 +258,27 @@ describe('given findDomainByDomain query', () => {
           findDomainByDomain: {
             dmarcSummaryByPeriod: {
               detailTables: {
-                fullPass: {
+                dmarcFailure: {
                   edges: [
                     {
-                      cursor: 'ZnVsbFBhc3M6MQ==',
+                      cursor: 'ZmFpbERtYXJjOjE=',
                       node: {
-                        dkimDomains: 'test.gc.ca',
-                        dkimSelectors: 'selector1',
-                        dnsHost: 'test.canada.ca',
-                        envelopeFrom: 'test.gc.ca',
-                        headerFrom: 'test.gc.ca',
-                        id: 'ZnVsbFBhc3M6MQ==',
+                        disposition: 'none',
+                        dkimDomains: '',
+                        dkimSelectors: '',
+                        dnsHost: 'test.dns.gc.ca',
+                        envelopeFrom: 'test.domain.canada.ca',
+                        headerFrom: 'test.domain.canada.ca',
+                        id: 'ZmFpbERtYXJjOjE=',
                         sourceIpAddress: '123.456.78.91',
-                        spfDomains: 'test.gc.ca',
-                        totalMessages: 50,
+                        spfDomains: 'test.domain.gc.ca',
+                        totalMessages: 30,
                       },
                     },
                   ],
                   pageInfo: {
-                    startCursor: 'ZnVsbFBhc3M6MQ==',
-                    endCursor: 'ZnVsbFBhc3M6MQ==',
+                    startCursor: 'ZmFpbERtYXJjOjE=',
+                    endCursor: 'ZmFpbERtYXJjOjE=',
                     hasNextPage: true,
                     hasPreviousPage: false,
                   },
